EES-5095 Move processor's `IDataSetService` into separate directory
